feat: add Sass module support and some styling examples
Closes #31
