+++ app/package.json
@@ -41,6 +41,7 @@
     "@testing-library/user-event": "^13.0.2",
     "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-prettier": "^4.0.0",
+    "node-sass": "^6.0.1",
     "prettier": "^2.2.1"
   }
 }
