+++ app/src/App/styles.module.scss
@@ -1,0 +1,14 @@
+.nav {
+  a[class="active"] {
+    background-color: darkorange;
+    color: white;
+    text-decoration: none;
+  }
+}
+
+.header {
+  p {
+    color: blue;
+    margin-top: 0;
+  }
+}
