+++ app/src/Tasks/index.js
@@ -1,6 +1,8 @@
 import * as React from "react";
 
-import * as apiClient from "./apiClient";
+import * as apiClient from "../apiClient";
+
+import styles from "./styles.module.scss";
 
 const Tasks = () => {
   const [tasks, setTasks] = React.useState([]);
@@ -21,7 +23,7 @@ const Tasks = () => {
 };
 
 const TaskList = ({ tasks }) => (
-  <ul>
+  <ul className={styles.list}>
     {tasks.map(({ id, name }) => (
       <li key={id}>{name}</li>
     ))}
@@ -47,7 +49,9 @@ const AddTask = ({ addTask }) => {
         New task:{" "}
         <input onChange={(e) => setTask(e.currentTarget.value)} value={task} />
       </label>
-      <button disabled={!canAdd}>Add</button>
+      <button disabled={!canAdd} className={styles.button}>
+        Add
+      </button>
     </form>
   );
 };
