+++ app/src/index.js
@@ -6,6 +6,8 @@ import { BrowserRouter as Router } from "react-router-dom";
 import App from "./App";
 import reportWebVitals from "./reportWebVitals";
 
+import "./global.scss";
+
 ReactDOM.render(
   <React.StrictMode>
     <Router>
